WIP - Major upgrade of streamlit app based to use our multipage structure
