+++ schema/dlt/EyeOnData.toml
@@ -1,0 +1,5 @@
+[datasets]
+dataset_path = "/Users/johnson30/git/LLNL/pEyeOn/schema/dlt/schemas"
+
+[app]
+page_title = "EyeOnData"
