+++ schema/dlt/pages/_base_page.py
@@ -1,0 +1,13 @@
+from abc import ABC, abstractmethod
+
+class BasePageLayout(ABC):
+
+    def __init__(self):
+        pass
+
+    @abstractmethod
+    def page_content(self):
+        pass
+
+    def display_page(self):
+        self.page_content()
