+++ schema/dlt/utils/config.py
@@ -1,0 +1,7 @@
+
+from dynaconf import Dynaconf
+
+settings = Dynaconf(
+    envvar_prefix='EyeOnData_',
+    settings_files=['EyeOnData.toml'],
+)
